# services/core/api/swagger/swagger.yaml
basePath: /api/v1
definitions:
  handler.errorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  handler.successResponse:
    properties:
      message:
        type: string
    type: object
  mapping.FileData:
    properties:
      barcode:
        type: string
      cost:
        type: integer
      title:
        type: string
    type: object
...
host: localhost:6001
info:
  contact: {}
  description: API Router
  title: Swagger Router API
  version: "1.0"
paths:
...
  /products/prices:
    delete:
      description: delete price files
      parameters:
      - description: file name of a price
        in: query
        name: filename
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete price files
      tags:
      - prices
    get:
      description: get all prices files by one product
      parameters:
      - description: barcode of product
        in: query
        name: barcode
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get all prices files by one product
      tags:
      - prices
    post:
      description: create price file
      parameters:
      - description: file data
        in: body
        name: filedata
        required: true
        schema:
          $ref: '#/definitions/mapping.FileData'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create price file
      tags:
      - prices
  /products/prices/{file}:
    get:
      description: get file of price by name
      parameters:
      - description: name of file
        in: path
        name: file
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get file of price by name
      tags:
      - prices
...
securityDefinitions:
  BearerAuth:
    description: 'Type ''Bearer TOKEN'' to correctly set the API Key. Example: Bearer
      <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"